Refactor RPC Methods logic (Fix network change) (#3341)
* Refactor RPC Methods

* Migrate to typescript

* Improve selectedAddress code

* Implement Swappable proxy for provider

* Fix url on rpc methods

* Fix app version code

* Temporarily fix Typescript issues on RPCMethodMiddleware

* Remove Beta
