+++ package.json
@@ -230,6 +230,7 @@
     "reselect": "^4.0.0",
     "rn-fetch-blob": "^0.12.0",
     "stream-browserify": "1.0.0",
+    "swappable-obj-proxy": "^1.1.0",
     "through2": "3.0.1",
     "unicode-confusables": "^0.1.1",
     "url": "0.11.0",
